+++ src/xcode/ENA/ENA/Source/HTTPClient/Resources/AvailableDays/__tests__/AvailableDaysResourceTests.swift
@@ -97,59 +97,5 @@ class AvailableDaysResourceTests: CWATestCase {
 		}
 		waitForExpectations(timeout: .medium)
 	}
-	
-	func testFetchDay_Success() throws {
-		let url = try XCTUnwrap(Bundle(for: type(of: self)).url(forResource: "api-response-day-2020-05-16", withExtension: nil))
-		let stack = MockNetworkStack(
-			httpStatus: 200,
-			headerFields: ["etAg": "\"SomeEtag\""],
-			responseData: try Data(contentsOf: url)
-		)
-
-		let successExpectation = expectation(
-			description: "expect error result"
-		)
-
-		let httpClient = HTTPClient.makeWith(mock: stack)
-		httpClient.fetchDay("2020-05-01", forCountry: "IT") { result in
-			defer { successExpectation.fulfill() }
-			switch result {
-			case let .success(sapPackage):
-				self.assertPackageFormat(for: sapPackage)
-			case let .failure(error):
-				XCTFail("a valid response should never yield and error like: \(error)")
-			}
-		}
-		waitForExpectations(timeout: .medium)
-	}
-
-	func testFetchDay_InvalidPackage() throws {
-		let stack = MockNetworkStack(
-			httpStatus: 200,
-			responseData: Data(bytes: [0xA, 0xB] as [UInt8], count: 2)
-		)
-
-		let successExpectation = expectation(
-			description: "expect error result"
-		)
-
-		let httpClient = HTTPClient.makeWith(mock: stack)
-		httpClient.fetchDay("2020-05-01", forCountry: "IT") { result in
-			defer { successExpectation.fulfill() }
-			switch result {
-			case .success:
-				XCTFail("An invalid server response should not result in success!")
-			case let .failure(error):
-				switch error {
-				case .invalidResponse:
-					break
-				default:
-					XCTFail("Incorrect error type \(error) received, expected .invalidResponse")
-				}
-			}
-		}
-		waitForExpectations(timeout: .medium)
-	}
-
 
 }
